major cleanup and refactor
