+++ analysis_options.yaml
@@ -9,9 +9,7 @@ analyzer:
 linter:
   rules:
     - always_declare_return_types
-#    - always_specify_types
     - annotate_overrides
-#    - avoid_as
     - avoid_empty_else
     - avoid_init_to_null
     - avoid_return_types_on_setters
@@ -42,7 +40,6 @@ linter:
     - package_prefixed_library_names
     - parameter_assignments
     - prefer_final_fields
-    - prefer_final_locals
     - prefer_is_not_empty
     - slash_for_doc_comments
     - sort_constructors_first
