+++ tool/build.dart
@@ -3,5 +3,5 @@ import 'package:build_runner/build_runner.dart';
 import 'phases.dart';
 
 main() async {
-  await build(xmlPhase, deleteFilesByDefault: true);
+  await build(jsonPhase, deleteFilesByDefault: true);
 }
