+++ tool/phases.dart
@@ -1,8 +1,7 @@
 import 'package:build_runner/build_runner.dart';
-import 'package:typewriter/builder/json_builder.dart';
-import 'package:typewriter/builder/xml_builder.dart';
+import 'package:typewriter/typewriter.dart';
 
-final jsonPhases = new PhaseGroup.singleAction(new JsonBuilder(),
+final jsonPhase = new PhaseGroup.singleAction(new JsonBuilder(),
     new InputSet('typewriter', const ['example/lib/codecs/codecs.dart']));
 
 final xmlPhase = new PhaseGroup.singleAction(new XmlBuilder(),
